🎨 問題消去を開発者用に変更

+++ src/app/api/quiz/delete-all/route.ts
@@ -1,7 +1,13 @@
 import connectToDatabase from "@/app/lib/conectMongoDB";
 import Question from "@/app/lib/models/quizModel";
+import { getServerSession } from "next-auth";
+import { NextResponse } from "next/server";
 
 export async function DELETE() {
+  const session = await getServerSession();
+  if (session?.user?.email != process.env.MY_EMAIL)
+    return NextResponse.json({ message: "権限がありません" }, { status: 500 });
+
   try {
     await connectToDatabase();
     await Question.deleteMany({});
